Add optional buffer to JSON scraper (#121)
Scraping logs from the stdout may be sometimes blocking for the
application that is sending logs because executor can not keep up with
sending logs due to a temporary network delays. The executor should
drop the logs if there is a risk that the writes to stdout will become
blocking.